# Remove variable_excluded_from_dispatch() check for factory functions.