--- a/aten/src/ATen/native/TensorFactories.cpp
+++ b/aten/src/ATen/native/TensorFactories.cpp
@@ -174,7 +174,6 @@ Tensor empty_cpu(IntArrayRef size, const TensorOptions& options_, c10::optional<
   TensorOptions options = options_.merge_in(TensorOptions().memory_format(optional_memory_format));
 
   AT_ASSERT(options.device().type() == DeviceType::CPU);
-  TORCH_INTERNAL_ASSERT(impl::variable_excluded_from_dispatch());
   check_size_nonnegative(size);
 
   c10::Allocator* allocator;