Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#46371 Remove variable_excluded_from_dispatch() check for factory functions.**

Differential Revision: [D24324545](https://our.internmc.facebook.com/intern/diff/D24324545)