@ailzhang 2020-10-15
Some perf data
## Before this PR
```
 λ ~ numactl -C 10 python repro.py
<torch.utils.benchmark.utils.common.Measurement object at 0x7f37e5560b50>
torch.empty((1,))
  Median: 2.26 us
  IQR:    0.07 us (2.22 to 2.30)
  442 measurements, 10000 runs per measurement, 1 thread
 λ ~ numactl -C 10 python repro.py
<torch.utils.benchmark.utils.common.Measurement object at 0x7f0f34bc6510>
torch.empty((1,))
  Median: 2.19 us
  IQR:    0.07 us (2.15 to 2.22)
  458 measurements, 10000 runs per measurement, 1 thread
```
## After this PR
```
 λ ~ numactl -C 10 python repro.py
<torch.utils.benchmark.utils.common.Measurement object at 0x7f9b7235ea50>
torch.empty((1,))
  Median: 1.96 us
  IQR:    0.05 us (1.93 to 1.98)
  511 measurements, 10000 runs per measurement, 1 thread
 λ ~ numactl -C 10 python repro.py
<torch.utils.benchmark.utils.common.Measurement object at 0x7f2b99b13410>
torch.empty((1,))
  Median: 2.01 us
  IQR:    0.06 us (1.98 to 2.04)
  498 measurements, 10000 runs per measurement, 1 thread
```

## Delta of instruction counts for torch.empty((1,))
```
1674 ???:0x000000000024f160
200 build/../torch/csrc/jit/frontend/tracer.h:torch::PythonArgs::intlist(int)
200 build/../c10/util/intrusive_ptr.h:torch::autograd::impl::set_pyobj(at::Tensor const&, _object*)
200 build/../c10/core/TensorImpl.cpp:c10::TensorImpl::autograd_meta() const
200 build/../c10/core/DefaultDtype.cpp:c10::get_default_dtype()
200 build/../aten/src/TH/THAllocator.cpp:getTHDefaultAllocator()
200 /usr/include/c++/7/tuple:c10::TensorImpl::TensorImpl(c10::Storage&&, c10::DispatchKeySet, caffe2::TypeMeta const&, c10::optional<c10::Device>)
200 /usr/include/c++/7/bits/atomic_base.h:at::Tensor at::detail::make_tensor<c10::TensorImpl, c10::intrusive_ptr<c10::StorageImpl, c10::detail::intrusive_target_default_null_type<c10::StorageImpl> >, c10::DispatchKey, caffe2::TypeMeta&>(c10::intrusive_ptr<c10::StorageImpl, c10::detail::intrusive_target_default_null_type<c10::StorageImpl> >&&, c10::DispatchKey&&, caffe2::TypeMeta&)
100 build/../c10/core/TensorOptions.h:at::native::empty_cpu(c10::ArrayRef<long>, c10::TensorOptions const&, c10::optional<c10::MemoryFormat>)
90 /build/glibc-OTsEL5/glibc-2.27/malloc/malloc.c:_int_free
45 /build/glibc-OTsEL5/glibc-2.27/malloc/malloc.c:_int_memalign
6 ???:0x0000000000215e40
-6 miniconda/include/python3.7m/pybind11/pytypes.h:pybind11::detail::function_call::~function_call()
-200 build/aten/src/ATen/core/TensorBody.h:at::native::empty_cpu(c10::ArrayRef<long>, c10::TensorOptions const&, c10::optional<c10::MemoryFormat>)
-200 build/../c10/util/typeid.h:c10::TensorOptions::TensorOptions()
-200 build/../c10/util/Optional.h:torch::utils::maybe_initialize_cuda(c10::TensorOptions const&)
-200 build/../c10/core/TensorImpl.h:torch::autograd::impl::set_pyobj(at::Tensor const&, _object*)
-200 build/../c10/core/DispatchKeySet.h:at::native::empty_cpu(c10::ArrayRef<long>, c10::TensorOptions const&, c10::optional<c10::MemoryFormat>)
-200 build/../aten/src/ATen/record_function.h:at::RecordFunction::RecordFunction(at::RecordScope)
-200 ???:0x0000000012bcaf0b
-200 /usr/include/c++/7/bits/stl_vector.h:torch::PythonArgParser::raw_parse(_object*, _object*, _object*, _object**)
-200 /usr/include/c++/7/bits/stl_vector.h:torch::PyWarningHandler::~PyWarningHandler()
-200 /usr/include/c++/7/bits/stl_construct.h:torch::PyWarningHandler::~PyWarningHandler()
-300 build/../c10/core/impl/LocalDispatchKeySet.h:c10::impl::tls_local_dispatch_key_set()
-300 build/../aten/src/ATen/native/TensorFactories.cpp:at::native::empty_cpu(c10::ArrayRef<long>, c10::TensorOptions const&, c10::optional<c10::MemoryFormat>)
-600 build/aten/src/ATen/BackendSelectRegister.cpp:at::(anonymous namespace)::empty_memory_format(c10::ArrayRef<long>, c10::TensorOptions const&, c10::optional<c10::MemoryFormat>)
-800 build/../c10/core/DispatchKey.cpp:c10::getAutogradKeyFromBackend(c10::DispatchKey)
-1700 ???:0x000000000024eaf0
-2800 build/../c10/core/impl/LocalDispatchKeySet.cpp:c10::impl::tls_local_dispatch_key_set()
-4400 /build/glibc-OTsEL5/glibc-2.27/elf/../sysdeps/x86_64/tls_get_addr.S:__tls_get_addr
```
Seems like an easy perf win if we just disable this check for factory ops. :D