@ailzhang 2020-10-15
questions from @bhosmer on the old thread: 
Just curious, how'd you arrive at the set of asserts to remove, just by spidering out from BackendSelect kernels? Asking bc I wouldn't have thought about the SparseTensor stuff as being affected by this change.

Curiosity question #2 - is anybody still asserting on variable_excluded_from_dispatch() after this change?

This is the minimal set I need to remove to make tests pass and I made sure they're called through factory ops. 
There're a few left in codebase, I think we can remove them in a separate PR. :D